+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ v3.1.0
 
 *Release date: In development*
 
+- Update `[CONTEXT:a.b.c]` replacement to allow lists access, e.g. [CONTEXT:list.1] to access the second element of list
+
 v3.0.2
 ------